api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
alembic==1.13.3
httpx==0.27.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pyotp==2.9.0
qrcode[pil]==7.4.2
anthropic==0.34.2
plaid-python==39.0.0
numpy==2.2.4
scikit-learn==1.6.1
joblib==1.4.2
openpyxl==3.1.5
matplotlib==3.9.2
reportlab==4.2.5
